requirements.txt
aio-pika==9.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiormq==6.7.7
aiosignal==1.3.2
aiosmtplib==2.0.2
alembic==1.12.0
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.28.0
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.19
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
ollama==0.4.7
openai==1.69.0
openai-whisper==20240930
orjson==3.10.16
packaging==24.2
pamqp==3.2.1
passlib==1.7.4
pika==1.3.2
propcache==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydub==0.25.1
python-dotenv==1.0.0
python-jose==3.3.0
python-ldap==3.4.3
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.23.2
Werkzeug==3.1.3
whisper==1.1.10
yarl==1.18.3
zstandard==0.23.0
